+++ sonicwall-firewall/release_notes.md
@@ -1,7 +1,16 @@
 #### Following enhancements have been made to the SonicWall Firewall connector in version 1.1.0:
-#### What's Improved
-- Added the following new actions and corresponding playbooks:
-    - Create Address Group Object Configuration
-
 #### What's Fix
 - Fixed an issue where an Unauthorized error occurred while executing connector actions.
+#### What's Improved
+- Rename the `Get Address Object Configuration` action to `Get Address Object`
+- Rename the `Create Address Object Configuration` action to `Create Address Object`
+- Rename the `Update Address Object Configuration` action to `Update Address Object`
+- Rename the `Delete Address Object Configuration` action to `Delete Address Object`
+
+- Added the following new actions and corresponding playbooks:
+    - Create Address Group
+    - Get Address Group
+    - Update Address in Group
+    - Delete Address From Group
+    - Add Address Object to Group
+    - Remove Address Object from Group